pyproject.toml
[project]
name = "minion-suite"
version = "0.1.0"
description = "AI agent suite — composable, vendor-agnostic agents starting with code review"
requires-python = ">=3.14"
dependencies = [
    "litellm>=1.50.0",
    "fastmcp>=0.4.0",
    "pydantic>=2.0",
    "aiosqlite>=0.20.0",
    "psycopg[binary]>=3.1",
    "nats-py>=2.9.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "python-gitlab>=4.0.0",
    "rich>=13.0",
]

[project.scripts]
minion = "minions.cli:main"

[tool.uv]
python-preference = "only-managed"

[tool.ruff]
line-length = 150
target-version = "py314"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "RUF",  # ruff-specific rules
]
ignore = [
    "E501",   # line length (handled by formatter)
    "SIM108", # ternary operator (project convention: no ternaries)
]

[tool.ruff.lint.isort]
known-first-party = ["minions"]
